backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt src/requirements/prod.in
#
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
    #   django-ninja-extra
cffi==1.17.1
    # via cryptography
contextlib2==21.6.0
    # via django-ninja-extra
cryptography==44.0.0
    # via
    #   django-ninja-jwt
    #   pyjwt
dj-database-url==2.3.0
    # via -r src/requirements/prod.in
django==4.2.23
    # via
    #   -r src/requirements/prod.in
    #   dj-database-url
    #   django-cors-headers
    #   django-ninja
    #   django-ninja-extra
    #   django-ninja-jwt
django-cors-headers==4.6.0
    # via -r src/requirements/prod.in
django-ninja==1.3.0
    # via
    #   -r src/requirements/prod.in
    #   django-ninja-extra
django-ninja-extra==0.22.3
    # via django-ninja-jwt
django-ninja-jwt[crypto]==5.3.5
    # via -r src/requirements/prod.in
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via pydantic
idna==3.10
    # via email-validator
injector==0.22.0
    # via django-ninja-extra
psycopg[binary]==3.2.4
    # via -r src/requirements/prod.in
psycopg-binary==3.2.4
    # via psycopg
pycparser==2.22
    # via cffi
pydantic[email]==2.10.6
    # via
    #   -r src/requirements/prod.in
    #   django-ninja
pydantic-core==2.27.2
    # via pydantic
pyjwt[crypto]==2.10.1
    # via django-ninja-jwt
python-decouple==3.8
    # via -r src/requirements/prod.in
sqlparse==0.5.3
    # via django
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   psycopg
    #   pydantic
    #   pydantic-core
